Jest setup for unit test
