+++ __test__/appointment.test.json
@@ -1,0 +1,19 @@
+[
+    {
+      "scheduleId": 4711,
+      "startDate": "2020-04-29",
+      "startTime": "10:00:00",
+      "endDate": "2020-04-29",
+      "endTime": "14:30:00",
+      "startBreak": "10:00:00",
+      "endBreak": "14:30:00",
+      "startBreak2": "00:00:00",
+      "endBreak2": "00:00:00",
+      "startBreak3": "00:00:00",
+      "endBreak3": "00:00:00",
+      "startBreak4": "00:00:00",
+      "endBreak4": "00:00:00",
+      "employeeId": 4712,
+      "employeeName": "John Doe"
+    }
+]
