+++ package.json
@@ -4,10 +4,12 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start:dev": "nodemon",
-    "start:prod": "node ./dist/app.js",
-    "build": "tsc"
+    "start:prod": "node -r ts-node/register/transpile-only -r tsconfig-paths/register ./dist/app.js",
+    "build": "tsc",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage"
   },
   "repository": {
     "type": "git",
