+++ tsconfig.json
@@ -102,6 +102,6 @@
     "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
   },
-  "exclude": ["node_modules", "dist"],
+  "exclude": ["node_modules", "dist", "__test__", "coverage"],
   "include": ["src"]
 }
